c1541: Make stepper motor control logic easier to read.
Swap "stp" lines: this way, the head moves to next higher track when
"stp" value increases, and to next lower track when "stp" value decreases.
Also, rename track_num to half_track, as that's what it's counting.
Fix mtr description. This is the spindle motor control, not the stepper
(head) motor control. This said, stepper motor is disabled when spindle
is not spinning.
Also, make "mode" comment clearer.