video_vicII_656x.vhd: Fix sprite priority.
If a lower-index background-priority (1) sprite's opaque pixel intersects
with a higher-index foreground-priority (0) sprite's opaque pixel, and the
screen bitmap has a foreground color at that same position, then the
screen bitmap's foreground color is visible.
This is counter-intuitive, as higher-index sprite's pixel alone would mask
the bitmap color, and adding another sprite top of it makes the bitmap
visible again, but this is how the VICII works.